Refactor SSL MiTM out of Connection
SSL MiTM logic has been crufting up connection handling, move all that
to a new internal data handler SslMitmHandler that has the setup logic.
